% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{glimpse_countsfile}
\alias{glimpse_countsfile}
\title{Glimpse into mutCaller counts.txt.gz file for the names of variants and number of reads}
\usage{
glimpse_countsfile(file)
}
\arguments{
\item{file}{path to file name (only 1)}
}
\value{
a table of counts for each variant contained in the mutCaller output file
}
\description{
See what variants are in a counts.txt.gz output file from mutCaller
}
